requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
Babel>=1.3 # BSD
neutron-lib>=0.0.1 # Apache-2.0
oslo.db>=4.1.0 # Apache-2.0
oslo.config>=3.7.0 # Apache-2.0
oslo.concurrency>=3.5.0 # Apache-2.0
oslo.log>=1.14.0 # Apache-2.0
oslo.messaging>=4.0.0 # Apache-2.0
oslo.service>=1.0.0 # Apache-2.0

